restructure assets and images folder
