+++ src/components/sponsor-card.js
@@ -1,18 +1,33 @@
 /** @jsx jsx */
 import { jsx, Card } from "theme-ui";
-import { DynamicImage } from "gatsby-plugin-image";
+import { StaticImage } from "gatsby-plugin-image";
 
-export default ({ id, companyName, imagePath, ...props }) => (
-  <Card
-    {...props}
-    sx={{
-      maxWidth: 256,
-    }}
-  >
-    CARD: {imagePath}
-    <img src={`../static/images/sponsors/${imagePath}`} />
-    <div sx={{ textAlign: "center" }}>
-      <h3 sx={{ color: "#2f2762" }}>{companyName}</h3>
-    </div>
-  </Card>
-);
+export default ({ id, companyName, imagePath, ...props }) => {
+  const sponsorQuery = graphql`
+    query ($imagePath: String) {
+      allFile(filter: { name: { eq: $imagePath } }) {
+        edges {
+          node {
+            childImageSharp {
+              fluid {
+                ...GatsbyImageSharpFluid
+              }
+            }
+          }
+        }
+      }
+    }
+  `;
+  return (
+    <Card
+      sx={{
+        maxWidth: 256,
+      }}
+    >
+      <img src={imagePath} />
+      <div sx={{ textAlign: "center" }}>
+        <h3 sx={{ color: "#2f2762" }}>{companyName}</h3>
+      </div>
+    </Card>
+  );
+};
