+++ src/misc/coc.mdx
@@ -3,7 +3,7 @@ title: "Code of Conduct"
 slug: "code-of-conduct"
 ---
 
-import bookImage from "../static/images/book.webp";
+import bookImage from "../images/book.webp";
 
 <div class="body-wrap">
    <section class="absolute w-screen -z-10 h-[80vh]">
