+++ src/misc/faq.mdx
@@ -3,7 +3,7 @@ title: "FAQ"
 slug: "FAQ"
 ---
 
-import bookImage from "../static/images/faq.webp";
+import bookImage from "../images/faq.webp";
 
 <div class="body-wrap">
    <section class="absolute w-screen -z-10 h-[80vh]">
