+++ src/pages/{mdx.frontmatter__slug}.js
@@ -2,11 +2,13 @@ import * as React from "react";
 import { graphql } from "gatsby";
 import { Header } from "../components/header";
 import Layout from "./../components/layout";
+import "../assets/stylesheets/application.scss";
 
 const MiscPage = ({ data, children }) => {
   return (
     <>
-      <Header classHeader={"header-misc"}/>
+      <Header classHeader={"header-misc"} />
+      {data}
       <Layout pageTitle="layout test">
         {children}
       </Layout>
